Exception handling - delete

+++ src/main/java/com/devdavicl/webservice/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.devdavicl.webservice.entities.User;
 import com.devdavicl.webservice.repositories.UserRepository;
+import com.devdavicl.webservice.services.exceptions.DatabaseException;
 import com.devdavicl.webservice.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,16 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+	    
+	        repository.deleteById(id);
+	    }
+	    catch (DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	public User update(Long id, User obj) {
